Treat the whole query as a template
This resolves an issue around SQL Lab queries that are using jinja
templating as they are wrapped and sent to the explore view using the
"Visualize" flow in SQL Lab. At the moment the template doesn't get
processed.

Note that this makes it possible to integrate templating in more places
like metric or column expressions.