DRAFT: notes on supporting Malloy in Superset
This PR is opened as a stratchbook to provide some ideas/pointers around what a
Superset / Malloy integration could look like.

In this PR you'll find a collection of pointers and notes, in-code, that highlight
some of the areas that are most-relevant in-regards to an eventual integration of Malloy.

Goal is to works towards an hackathon-type approach that would
help identify what needs to be done in order to properly
support Malloy.

Eventually progress here and elsewhere will inform a Superset Improvement Proposal with the
key findings from this process.

NOTE: from an information architecture standpoint, different approaches
are possible. For the hacking/feasibility-study side, I think it makes
sense to **bend** existing entities/objects to support malloy in this way:

- Think of "Database Connections" as a "Connections" and fill in the
  attributes of a "Malloy Connection" to connect to a "Malloy Publisher"
  instance, or a "GitHub-hosted" mallow project (?) or even an
  MS2-hosted construct (?)
- Think of the `Dataset` entity as a "Data Source", and create a new
  "Data Source" type for a "Malloy Project" right next to "Physical
  Dataset" and "Virtual Dataset". When picking a "Malloy Project",
  we'd interogate the connection to see which Project to use as a
  source. There's probably a need to "sync project" to fetch the metrics
  and dimensions as Superset assumes a local representation of
  metrics/dims. When syncing the project, we fetch and materialize the
  metadata locally, so that they can be represented in the same way
  as dimensions/metrics on our side
- Use and abuse the `superset/db_engine_specs` module as a way to gather
  all the malloy-specific logic outside the core codebase. We may need to
  refactor some things from the core codebase and onto that package

Potentially - and that would take more time/effort - it probably make sense to
introduce new abstractions in the right places. More specifically formalizing the
distinction between what's an actual "SQL Dataset" from a more proper
"Semantic Layer" data source. This may require a fair amount of
refactoring, formalizing a proper "Data Source" interface on top of both
"SQL-datasets" and "Semantic Layer".

Ideating around abstractions, it could make sense to go towards this hierarchy of abstractions:

- Datasource: anything that can be used for data exploration, in essence
  exposes metrics/dimensions. Exposes metadata, can run dimensional
  queries and return dataframes, ...
  - BaseSemanticLAyer: base to all semantic layer
    - MalloySemanticLayer: first/core implementation
    - DbtSemanticLayer: eventually could support this within this
      structure
    - SqlMeshSemanticLayer: same here, could support if/when it comes out
  - Dataset: current code logic for all SqlAlchemy-powered dataset, supports both physical and virtual datasets
  - Other types of datasources? Going off rails here but: GoogleAnalytics, MicroTragedy, Business Objects, anything that can expose metrics/dims and execute dimensional queries.